Enhance styling to improve visibility and clarity
This affects the "helpful" and "decline" buttons for individual flags,
as well as the decline reason buttons.

Bold emphasis was added to key phrases in the custom reasons
in order to make them easier to pick out at a glance.
The last-used reason was also indicated clearly in the UI.